refactor: move pl specifics to pl::extra::expr
